계약 조회, 삭제 기능 재검토, 테스트 실패

+++ orderplanning/src/main/java/com/overc1ock/domain/ContractVO.java
@@ -15,7 +15,7 @@ public class ContractVO {
 	private String subcontractor_email;
 	private String subcontractor_person;
 	private String product_name;
-	private Integer item_code;
+	private Long item_code;
 	private String item_name;
 	private String standard;
 	private String material;
